requirements-dev.txt
# Development tools
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
pytest-mock==3.12.0
pytest-benchmark==4.0.0
testcontainers==3.7.1
redis==6.2.0
pre-commit==3.6.0

# Security scanning
bandit==1.7.6
safety==3.0.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Type stubs
types-requests==2.31.0.20240106
types-pyyaml==6.0.12.12

# OpenTelemetry instrumentation packages
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-sqlalchemy==0.43b0
opentelemetry-instrumentation-asyncpg==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-instrumentation-aiohttp-client==0.43b0
opentelemetry-instrumentation-psycopg2==0.43b0
opentelemetry-instrumentation-urllib3==0.43b0
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-instrumentation-jinja2==0.43b0

# Additional exporters
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-exporter-otlp==1.22.0

# Prometheus integration
opentelemetry-exporter-prometheus-remote-write==0.43b0
